Put Settings class into it's own file
